feat(core): update client v2 to explicitely parse transaction response
